requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
boto3==1.9.199            # via django-s3-storage
botocore==1.12.199        # via boto3, s3transfer
certifi==2019.6.16        # via requests
chardet==3.0.4            # via requests
click==7.0                # via pip-tools
coreapi==2.3.3            # via drf-yasg
coreschema==0.0.4         # via coreapi, drf-yasg
django-cors-headers==2.5.2
django-environ==0.4.5
django-filter==2.1.0
django-s3-storage==0.12.4
django==2.2.4
djangorestframework==3.9.4
docutils==0.14            # via botocore
drf-yasg==1.16.1
idna==2.8                 # via requests
inflection==0.3.1         # via drf-yasg
itypes==1.1.0             # via coreapi
jinja2==2.10.1            # via coreschema
jmespath==0.9.4           # via boto3, botocore
markupsafe==1.1.1         # via jinja2
packaging==19.1           # via drf-yasg
pip-tools==4.0.0
psycopg2==2.8.2
python-dateutil==2.8.0    # via botocore
pytz==2019.2              # via django
requests==2.22.0
ruamel.yaml.clib==0.1.0   # via ruamel.yaml
ruamel.yaml==0.15.100     # via drf-yasg
s3transfer==0.2.1         # via boto3
six==1.12.0               # via drf-yasg, pip-tools, python-dateutil
sqlparse==0.3.0           # via django
uritemplate==3.0.0        # via coreapi, drf-yasg
urllib3==1.25.3           # via botocore, requests
uwsgi==2.0.18
